services/embedding_service/requirements.txt
Flask>=3.0.2
vertexai>=0.0.1
google-api-core>=2.11.0
google-cloud-core>=2.4.1
google-cloud-aiplatform>=1.40.0
google-cloud-storage>=2.14.0
unstructured[all-docs]>=0.12.3
pinecone-client>=3.0.2
python-dotenv>=1.0.1

gunicorn>=21.2.0
